+++ pkg/local_object_storage/blobstor/common/storage_test.go
@@ -15,6 +15,16 @@ import (
 )
 
 func TestCopy(t *testing.T) {
+	testCopy(t, common.Copy)
+}
+
+func TestCopyBatched(t *testing.T) {
+	testCopy(t, func(dst, src common.Storage) error {
+		return common.CopyBatched(dst, src, 7)
+	})
+}
+
+func testCopy(t *testing.T, copier func(dst, src common.Storage) error) {
 	dir := t.TempDir()
 	const nObjects = 100
 
@@ -39,17 +49,18 @@ func TestCopy(t *testing.T) {
 
 	dst := peapod.New(filepath.Join(dir, "peapod.db"), 0o600, 10*time.Millisecond)
 
-	err := common.Copy(dst, src)
+	err := copier(dst, src)
 	require.NoError(t, err)
 
 	require.NoError(t, dst.Open(true))
 	t.Cleanup(func() { _ = dst.Close() })
 
+	dstObjs := make(map[oid.Address][]byte, nObjects)
+
 	err = dst.Iterate(func(addr oid.Address, data []byte, _ []byte) error {
-		origData, ok := mObjs[addr]
-		require.True(t, ok)
-		require.Equal(t, origData, data)
+		dstObjs[addr] = data
 		return nil
 	}, nil)
+	require.Equal(t, mObjs, dstObjs)
 	require.NoError(t, err)
 }